+++ src/main/kotlin/io/ioNeo4j/VertexNeo4j.kt
@@ -7,17 +7,16 @@ import org.springframework.data.neo4j.core.schema.Relationship.Direction
 class VertexNeo4j(
     @Id @GeneratedValue
     var id: Long? = null,
+
     @Property
-    var label: String = "",
+    val x: Double? = null,
     @Property
-    val x: Float? = null,
-    @Property
-    val y: Float? = null,
+    val y: Double? = null,
     @Property
     val isDirected: Boolean = false,
     @Property
     val isWeighted: Boolean = false,
 ) {
     @Relationship(type = "Edge", direction = Direction.OUTGOING)
-    var edges: MutableList<EdgeNeo4j> = mutableListOf()
+    var edges: List<EdgeNeo4j>? = null
 }
