--- src/main/kotlin/ioNeo4j/EdgeNeo4j.kt
@@ -1,16 +1,0 @@
-package ioNeo4j
-
-import org.springframework.data.neo4j.core.schema.*
-
-@RelationshipProperties
-class EdgeNeo4j(
-    _vertex: VertexNeo4j,
-    @GeneratedValue
-    @RelationshipId
-    private var id: Long? = null
-) {
-    private var weight: Long? = null;
-
-    @TargetNode
-    private var vertex = _vertex;
-}
