+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.neo4j.authentication.username=neo4j
 spring.neo4j.authentication.password=di03ma01mot07
 logging.level.org.neo4j.driver=DEBUG
 logging.level.org.springframework.data.neo4j=DEBUG
+spring.main.allow-bean-definition-overriding=true
